+++ src/Control/Lens/TH.hs
@@ -37,8 +37,9 @@ import Control.Lens.Fold
 import Control.Lens.Getter
 import Control.Lens.Iso
 import Control.Lens.Setter
-import Control.Lens.Type
+import Control.Lens.Tuple
 import Control.Lens.Traversal
+import Control.Lens.Type
 import Control.Monad
 import Data.Char (toLower)
 import Data.Foldable