+++ src/Data/IntSet/Lens.hs
@@ -17,6 +17,7 @@ module Data.IntSet.Lens
 
 import Control.Applicative
 import Control.Lens
+import Control.Lens.Internal
 import Data.IntSet as IntSet
 
 -- | This 'Lens' can be used to read, write or delete a member of an 'IntSet'