+++ src/Data/Set/Lens.hs
@@ -15,9 +15,10 @@ module Data.Set.Lens
   ) where
 
 import Control.Applicative
-import Control.Lens.Type
-import Control.Lens.Setter
 import Control.Lens.Getter
+import Control.Lens.Internal
+import Control.Lens.Setter
+import Control.Lens.Type
 import Data.Set as Set
 
 -- | This 'Lens' can be used to read, write or delete a member of a 'Set'