+++ CHANGELOG.markdown
@@ -1,6 +1,7 @@
 2.4
 -----
 * Added the indexed `Kleene` store to `Control.Lens.Internal`
+* Moved `Gettable`, `Accessor`, `Settable` and `Mutator` to `Control.Lens.Internal`
 * Added `cloneTraversal` to `Control.Lens.Traversal`
 * Renamed `clone` to `cloneLens` in `Control.Lens.Type`
 * Generalized the type of `zoom` to subsume `focus`.